git-svn-id: http://svn.bluexml.com/private/S-IDE@13593 c4e9bb49-730a-0410-b52d-bc9a0859fdfe

+++ Integration/trunk/com.bluexml.side.Integration.buildHudson/src/com/bluexml/side/integration/buildHudson/utils/Utils.java
@@ -1700,8 +1700,8 @@ public class Utils {
 			if (!new File(getFinalDirectory()).exists())
 				new File(getFinalDirectory()).mkdir();
 
-			System.out.println("\t- Suppression de l'ancien update site" );
-			/*if (new File(getUpdateSiteDir() + File.separator + getCodeName()
+			/*System.out.println("\t- Suppression de l'ancien update site" );
+			if (new File(getUpdateSiteDir() + File.separator + getCodeName()
 					+ File.separator + "features").exists()) {
 				FileHelper.deleteFile(new File(getUpdateSiteDir() + File.separator
 						+ getCodeName() + File.separator + "features"));
